pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "social-downloader"
version = "0.1.0"
description = "Social media video downloader CLI"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "yt-dlp>=2023.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "telethon>=1.30.0",
    "requests>=2.28.0",
    "httpx>=0.24.0",
    "kmp>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
social = "social.cli.app:app"
social-download = "social.cli.app:app"

[tool.setuptools]
packages = ["social", "social.cli", "social.cli.commands", "social.core", "social.platforms", "social.services", "social.bot", "social.bot.handlers"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
markers = [
    "e2e: End-to-end tests with real integration",
    "slow: Slow running tests",
    "unit: Unit tests",
    "integration: Integration tests",
    "download: Download tests",
]
